fixed #13411 - [Task]: Remove Elasticsearch 7 support
