+++ bundles/EcommerceFrameworkBundle/src/IndexService/ProductList/ElasticSearch/DefaultElasticSearch8.php
@@ -19,44 +19,5 @@ use Elastic\Elasticsearch\Client;
 
 class DefaultElasticSearch8 extends AbstractElasticSearch
 {
-    /**
-     * send a request to elasticsearch
-     */
-    protected function sendRequest(array $params): array
-    {
-        /**
-         * @var Client $esClient
-         */
-        $esClient = $this->tenantConfig->getTenantWorker()->getElasticSearchClient();
-        $result = [];
 
-        if ($esClient instanceof Client) {
-            if ($this->doScrollRequest) {
-                $params = array_merge(['scroll' => $this->scrollRequestKeepAlive], $params);
-                //kind of dirty hack :/
-                $params['body']['size'] = $this->getLimit();
-            }
-
-            $result = $esClient->search($params)->asArray();
-
-            if ($this->doScrollRequest) {
-                $additionalHits = [];
-                $scrollId = $result['_scroll_id'];
-
-                while (true) {
-                    $additionalResult = $esClient->scroll(['scroll_id' => $scrollId, 'scroll' => $this->scrollRequestKeepAlive])->asArray();
-
-                    if (count($additionalResult['hits']['hits'])) {
-                        $additionalHits = array_merge($additionalHits, $additionalResult['hits']['hits']);
-                        $scrollId = $additionalResult['_scroll_id'];
-                    } else {
-                        break;
-                    }
-                }
-                $result['hits']['hits'] = array_merge($result['hits']['hits'], $additionalHits);
-            }
-        }
-
-        return $result;
-    }
 }
